pyproject.toml
[tool.poetry]
name = "backlogmcp"
version = "0.1.0"
description = "BacklogMCPは、Backlog SaaSをModel Context Protocol (MCP)経由で操作できるようにするプロジェクトです。"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.12"
uvicorn = {extras = ["standard"], version = "^0.34.2"}
pydantic = "^2.11.3"
httpx = "^0.28.1"
python-dotenv = "^1.1.0"
fastapi-mcp = "0.3.3"
gunicorn = "^23.0.0"
mangum = "^0.19.0"
aws-xray-sdk = "^2.14.0"
uv = "^0.6.17"
requests = "^2.32.2"
six = "^1.17.0"
mcp = "^1.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--cov=app --cov-report=term --cov-report=html"

[tool.black]
line-length = 88
target-version = ["py310"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.poetry.scripts]
start = "app.main:start"
test = "scripts.run_tests:main"
setup-dev = "scripts.setup_dev:main"
